requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
Django==4.1.7
django-ckeditor==6.5.1
django-environ==0.10.0
django-js-asset==2.0.0
fontawesomefree==6.4.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-decouple==3.8
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
